+++ sphinx/screens-detail/support-ticket/support-ticket/src/main/java/chat/sphinx/support_ticket/ui/SupportTicketFragment.kt
@@ -81,14 +81,23 @@ internal class SupportTicketFragment: SideEffectFragment<
             @Exhaustive
             when(viewState) {
                 is SupportTicketViewState.Empty -> {
-                    includeSupportTicketLogText.textViewSupportTicketLogs.text = ""
+                    logsDivider.gone
+                    includeSupportTicketLogText.root.gone
+                    includeSupportTicketLogApp.root.gone
                     progressBarSupportTicketLogs.gone
                 }
                 is SupportTicketViewState.Fetched -> {
+                    logsDivider.visible
+                    includeSupportTicketLogText.root.visible
+                    includeSupportTicketLogApp.root.visible
                     includeSupportTicketLogText.textViewSupportTicketLogs.text = viewState.logs
+                    includeSupportTicketLogApp.textViewSupportTicketLogs.text = viewState.appLogs
                     progressBarSupportTicketLogs.gone
                 }
-                SupportTicketViewState.LoadingLogs -> {
+                is SupportTicketViewState.LoadingLogs -> {
+                    logsDivider.gone
+                    includeSupportTicketLogText.root.gone
+                    includeSupportTicketLogApp.root.gone
                     progressBarSupportTicketLogs.visible
                 }
             }