+++ sphinx/screens-detail/support-ticket/support-ticket/src/main/java/chat/sphinx/support_ticket/ui/SupportTicketViewState.kt
@@ -7,5 +7,8 @@ internal sealed class SupportTicketViewState: ViewState<SupportTicketViewState>(
 
     object LoadingLogs: SupportTicketViewState()
 
-    data class Fetched(val logs: String): SupportTicketViewState()
+    data class Fetched(
+        val logs: String,
+        val appLogs: String
+    ): SupportTicketViewState()
 }