+++ sphinx/screens/dashboard/dashboard/src/main/java/chat/sphinx/dashboard/ui/DashboardFragment.kt
@@ -2,6 +2,7 @@ package chat.sphinx.dashboard.ui
 
 import android.content.Context
 import android.os.Bundle
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.widget.ImageView
@@ -70,6 +71,8 @@ internal class DashboardFragment : MotionLayoutFragment<
 
     private val podcastPlayerBinding: LayoutPodcastPlayerFooterBinding
         get() = binding.layoutPodcastPlayerFooter
+
+    var timeTrackerStart: Long = 0
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -260,17 +263,21 @@ internal class DashboardFragment : MotionLayoutFragment<
         }
     }
 
+    private fun onPodcastBarDismissed() {
+        dashboardPodcastViewModel.forcePodcastStop()
+        dashboardPodcastViewModel.trackPodcastConsumed()
+
+        podcastPlayerBinding.root.gone
+        binding.swipeRevealLayoutPlayer.gone
+        binding.imageViewBottomBarShadow.visible
+    }
+
     private fun setupPodcastPlayerFooter() {
         binding.swipeRevealLayoutPlayer.setSwipeListener(object: SwipeRevealLayout.SwipeListener {
             override fun onClosed(view: SwipeRevealLayout?) {}
 
             override fun onOpened(view: SwipeRevealLayout?) {
-                dashboardPodcastViewModel.forcePodcastStop()
-                dashboardPodcastViewModel.trackPodcastConsumed()
-
-                podcastPlayerBinding.root.gone
-                binding.swipeRevealLayoutPlayer.gone
-                binding.imageViewBottomBarShadow.visible
+                onPodcastBarDismissed()
             }
 
             override fun onSlide(view: SwipeRevealLayout?, slideOffset: Float) {}
@@ -429,6 +436,7 @@ internal class DashboardFragment : MotionLayoutFragment<
                         is LoadResponse.Loading -> {
                             dashboardHeader.progressBarDashboardHeaderNetwork.invisibleIfFalse(true)
                             dashboardHeader.textViewDashboardHeaderNetwork.invisibleIfFalse(false)
+                            timeTrackerStart = System.currentTimeMillis()
                         }
                         is Response.Error -> {
                             dashboardHeader.progressBarDashboardHeaderNetwork.invisibleIfFalse(false)
@@ -449,6 +457,8 @@ internal class DashboardFragment : MotionLayoutFragment<
                                     R.color.primaryGreen
                                 )
                             )
+                            Log.d("TimeTracker", "Your node went online in ${System.currentTimeMillis() - timeTrackerStart} milliseconds")
+                            viewModel.sendAppLog("- Your node went online in ${System.currentTimeMillis() - timeTrackerStart} milliseconds")
                         }
                     }
                 }