+++ sphinx/screens/dashboard/dashboard/src/main/java/chat/sphinx/dashboard/ui/feed/all/FeedAllSideEffect.kt
@@ -1,18 +1,16 @@
 package chat.sphinx.dashboard.ui.feed.all
 
-import android.content.Context
+import androidx.fragment.app.FragmentActivity
+import chat.sphinx.dashboard.R
 import chat.sphinx.resources.SphinxToastUtils
 import io.matthewnelson.android_feature_toast_utils.show
 import io.matthewnelson.concept_views.sideeffect.SideEffect
 
-sealed class FeedAllSideEffect: SideEffect<Context>() {
+sealed class FeedAllSideEffect: SideEffect<FragmentActivity>() {
 
-    class Notify(
-        private val msg: String,
-        private val notificationLengthLong: Boolean = true
-    ): FeedAllSideEffect() {
-        override suspend fun execute(value: Context) {
-            SphinxToastUtils(toastLengthLong = notificationLengthLong).show(value, msg)
+    object RefreshWhilePlaying: FeedAllSideEffect() {
+        override suspend fun execute(value: FragmentActivity) {
+            SphinxToastUtils(false).show(value, value.getString(R.string.refresh_while_loading))
         }
     }
 }