+++ sphinx/screens/splash/splash/src/main/java/chat/sphinx/splash/ui/SplashViewModel.kt
@@ -1,9 +1,11 @@
 package chat.sphinx.splash.ui
 
+import android.util.Log
 import androidx.lifecycle.viewModelScope
 import app.cash.exhaustive.Exhaustive
 import chat.sphinx.concept_background_login.BackgroundLoginHandler
 import chat.sphinx.concept_relay.RelayDataHandler
+import chat.sphinx.concept_repository_actions.ActionsRepository
 import chat.sphinx.concept_repository_lightning.LightningRepository
 import chat.sphinx.onboard_common.OnBoardStepHandler
 import chat.sphinx.onboard_common.model.OnBoardStep
@@ -18,19 +20,20 @@ import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.firstOrNull
 import kotlinx.coroutines.launch
 import javax.inject.Inject
-
 @HiltViewModel
 internal class SplashViewModel @Inject constructor(
     private val authenticationCoordinator: AuthenticationCoordinator,
     private val backgroundLoginHandler: BackgroundLoginHandler,
     dispatchers: CoroutineDispatchers,
     private val lightningRepository: LightningRepository,
+    private val actionsRepository: ActionsRepository,
     private val navigator: SplashNavigator,
     private val onBoardStepHandler: OnBoardStepHandler,
 ): BaseViewModel<
         SplashViewState
         >(dispatchers, SplashViewState.Idle)
 {
+    private val timeTrackerStart = System.currentTimeMillis()
 
     private var screenInit: Boolean = false
     fun screenInit() {
@@ -105,6 +108,9 @@ internal class SplashViewModel @Inject constructor(
                                     }
                                     null -> {
                                         navigator.toDashboardScreen(updateBackgroundLoginTime = true)
+
+                                        Log.d("TimeTracker", "Dashboard screen was call in ${System.currentTimeMillis() - timeTrackerStart} milliseconds")
+                                        actionsRepository.setAppLog("- Dashboard screen was call in ${System.currentTimeMillis() - timeTrackerStart} milliseconds")
                                     }
                                 }
                             }
@@ -140,4 +146,6 @@ internal class SplashViewModel @Inject constructor(
             }
         }
     }
+
 }
+